+++ src/WebJobs.Script/ScriptConstants.cs
@@ -165,6 +165,7 @@ namespace Microsoft.Azure.WebJobs.Script
         public const string ExtensionBundleTemplatesFile = "templates.json";
         public const string ExtensionBundleResourcesFile = "Resources.json";
         public const string DefaultExtensionBundleId = "Microsoft.Azure.Functions.ExtensionBundle";
+        public const string WorkFlowExtensionBundleId = "Microsoft.Azure.Functions.ExtensionBundle.Workflows";
         public const string ExtensionBundleForAppServiceWindows = "win-any";
         public const string ExtensionBundleForAppServiceLinux = "linux-x64";
         public const string ExtensionBundleForNonAppServiceEnvironment = "any-any";